remove: 불필요한 props drilling으로 인해 파일 분리 삭제
